+++ CHANGELOG.md
@@ -12,6 +12,7 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Added static method Uri::fromString() to get UriInterface from string.
 - Added LoggerInterface proxy methods for each priority.
 - Added RouterInterface::assemble() parameter $keepUnresolved to indicate that a missing path parameter must be ignored.
+- Added ReflectionResolver::getServiceFromClosure() method to use a closure as a class wrapper.
 
 ### Changed
 
@@ -22,7 +23,6 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/) a
 - Replaced STDIN and STDOUT with fopen() function and PHP_SAPI constant.
 - Changed LogInterface::getDateTime() return type from DateTime to DateTimeImmutable.
 - Bumped PHP version from 8.2 to 8.3.
-- Changed ReflectionResolver::addReflection() method to accept classname and closure.
 
 ### Deprecated